Redo up and down annotations
The new versions no longer use MList and Deleteable. This makes them
easier to use with lenses and makes the code simplier.

'query' is now a traversal and the old 'lenses' are also traversals.
The old lenses wheren't  quite valid before because of the empty diagram
case. If we want to consider the empty diagram and a diagram with an
empty leafU node the same thing, they could be changed back to lenses.