Add foldDia' variant
For access to styles that can affect multiple parts of a diagram (like
clippings).

Annotations also now have b and v types. This is so we can support
'fading' in the future. Fading is using the luminance of a diagram as
the alpha channel of everything below the annotation. Similarly to group
opacity, this needs to be an annotation, but it is also affected by
transforms.